+++ src/os2datascanner/projects/admin/organizations/tests/test_org_views.py
@@ -40,6 +40,7 @@ class TestOrganizationListViews:
     def test_administrator_for_list(self, client, user_admin, test_client, other_client):
         """Administrators should only be able to see the client, that they are
         administrator for."""
+
         client.force_login(user_admin)
 
         url = reverse_lazy("organization-list")
@@ -462,6 +463,9 @@ class TestOrganizationalUnitListView:
         """Administrators should be able to see the units belonging to
         organizations, belonging to clients, for which they are
         administrators."""
+        # Add user permission
+        user_admin.user_permissions.add(Permission.objects.get(codename="view_organizationalunit"))
+
         client.force_login(user_admin)
 
         # Add accounts to ous
@@ -487,6 +491,9 @@ class TestOrganizationalUnitListView:
 
     def test_regular_user_list(self, client, user, test_org, test_org2):
         """Users with no priviliges should not be able to see any units."""
+        # Add user permission
+        user.user_permissions.add(Permission.objects.get(codename="view_organizationalunit"))
+
         client.force_login(user)
 
         # URL to all units from organization 1
@@ -575,6 +582,9 @@ class TestOrganizationalUnitListViewAddRemoveManagers:
             olsen_banden):
         """An administrator should be able to add managers to units they are
         administrator for."""
+        # Add user permission
+        user_admin.user_permissions.add(Permission.objects.get(codename="view_organizationalunit"))
+
         client.force_login(user_admin)
 
         # URL to all units from organization 1
@@ -604,6 +614,9 @@ class TestOrganizationalUnitListViewAddRemoveManagers:
             olsen_banden):
         """An administrator should be able to remove managers from the units
         they are administrators for."""
+        # Add user permission
+        user_admin.user_permissions.add(Permission.objects.get(codename="view_organizationalunit"))
+
         client.force_login(user_admin)
 
         Position.managers.create(account=fritz, unit=nisserne)
@@ -636,6 +649,9 @@ class TestOrganizationalUnitListViewAddRemoveManagers:
             olsen_banden):
         """An unprivileged user should not be able to add managers to any
         units."""
+        # Add user permission
+        user.user_permissions.add(Permission.objects.get(codename="view_organizationalunit"))
+
         client.force_login(user)
 
         # URL to all units from organization 1
@@ -665,6 +681,9 @@ class TestOrganizationalUnitListViewAddRemoveManagers:
             olsen_banden):
         """An unprivileged user should not be able to remove managers from
         any units."""
+        # Add user permission
+        user.user_permissions.add(Permission.objects.get(codename="view_organizationalunit"))
+
         client.force_login(user)
 
         Position.managers.create(account=fritz, unit=nisserne)
@@ -684,3 +703,53 @@ class TestOrganizationalUnitListViewAddRemoveManagers:
         assert response2.status_code == 404
         assert nisserne in fritz.get_managed_units()
         assert olsen_banden in egon.get_managed_units()
+
+
+@pytest.mark.django_db
+class TestOrganizationalUnitVisibilityView:
+
+    def test_orgunit_visibility_access_as_superuser(self, client, superuser, test_org):
+        """Superusers should have access to the view."""
+        client.force_login(superuser)
+        url = reverse_lazy("edit-orgunit-visibility-view", kwargs={"org_slug": test_org.slug})
+        response = client.get(url)
+
+        assert response.status_code == 200
+
+    def test_orgunit_visibility_access_as_administrator(self, client, user_admin, test_org):
+        """Superusers should have access to the view."""
+        user_admin.user_permissions.add(
+                Permission.objects.get(codename="change_visibility_organizationalunit"))
+        client.force_login(user_admin)
+        url = reverse_lazy("edit-orgunit-visibility-view", kwargs={"org_slug": test_org.slug})
+        response = client.get(url)
+
+        assert response.status_code == 200
+
+    def test_orgunit_visibility_access_as_user(self, client, user, test_org):
+        """Superusers should have access to the view."""
+        user.user_permissions.add(
+                Permission.objects.get(codename="change_visibility_organizationalunit"))
+        client.force_login(user)
+        url = reverse_lazy("edit-orgunit-visibility-view", kwargs={"org_slug": test_org.slug})
+        response = client.get(url)
+
+        assert response.status_code == 404
+
+    @pytest.mark.parametrize("has_perm", [True, False])
+    def test_orgunit_visibility_access_permission(self, client, user_admin, has_perm, test_org):
+        if has_perm:
+            user_admin.user_permissions.add(
+                Permission.objects.get(codename="change_visibility_organizationalunit"))
+        client.force_login(user_admin)
+
+        url = reverse_lazy("edit-orgunit-visibility-view", kwargs={"org_slug": test_org.slug})
+
+        response = client.get(url)
+
+        if has_perm:
+            assert response.status_code == 200
+        else:
+            assert response.status_code == 403
+
+    # TODO: Write tests for hiding and un-hiding OUs: #63234